third_party/WebKit/LayoutTests/http/tests/htmlimports/import-and-script-module-external.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel="import" href="resources/module-script-external.html">
<script>
test(() => {
  assert_true(window.runExternalScript);
}, "External module script ran in HTML import.");
</script>
